requirements.txt
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
psycopg2==2.7.7
dj-rest-auth==2.1.1
django-allauth>=0.44
